fix: prevent app rerenders from NavigationBar logic
