+++ app/App.js
@@ -5,15 +5,16 @@
 
 import { useFonts } from 'expo-font';
 import { Provider } from 'react-redux';
-import React, { useEffect } from 'react';
+import React from 'react';
 import { useKeepAwake } from 'expo-keep-awake';
 import * as SplashScreen from 'expo-splash-screen';
 import * as NavigationBar from "expo-navigation-bar";
-import { StatusBar, setStatusBarHidden } from 'expo-status-bar';
+import { setStatusBarHidden } from 'expo-status-bar';
 
 import store from './src/state';
 
 import AppActual from './src/AppActual';
+import AppShellStyles from './src/components/AppShellStyles';
 
 // Show splash screen until fonts are loaded
 SplashScreen.preventAutoHideAsync();
@@ -22,27 +23,13 @@ SplashScreen.preventAutoHideAsync();
 NavigationBar.setPositionAsync("absolute");
 NavigationBar.setVisibilityAsync("hidden");
 NavigationBar.setBehaviorAsync("inset-swipe");
+NavigationBar.setButtonStyleAsync("light");
 NavigationBar.setBackgroundColorAsync("#00000080"); // `rgba(0,0,0,0.5)`
 setStatusBarHidden(true, "fade");
 
 export default function App() {
 	// Keep screen awake
 	useKeepAwake();
-
-	const visibility = NavigationBar.useVisibility()
-
-	// Hide navigation bar after a few seconds
-	useEffect(() => {
-		if (visibility !== "visible") return;
-
-		const interval = setTimeout(() => {
-			NavigationBar.setVisibilityAsync("hidden");
-		}, /* 3 Seconds */ 3000);
-
-		return () => {
-			clearTimeout(interval);
-		};
-	}, [visibility]);
 
 	// Load fonts
 	let [fontsLoaded] = useFonts({
@@ -54,13 +41,11 @@ export default function App() {
 	else return null;
 
 	return (
-		<Provider store={store}>
-			<StatusBar
-				hidden={true}
-				translucent
-				backgroundColor="transparent"
-			/>
-			<AppActual />
-		</Provider>
+		<>
+			<AppShellStyles />
+			<Provider store={store}>
+				<AppActual />
+			</Provider>
+		</>
 	);
 }
